Created directories for each of the examples.
The "boring" example is the most heavily reworked in this commit
in order to do the first thornydev blog entry on it.

util now has go and go& macros.  generator-tq is the primary
generator example using a Java TransferQueue.  The function for
that was changed to (go-channel in the util lib).

No tests yet.  README updated for public consumption.

The intro.md is just my notes to date.